+++ src/data/data.h
@@ -7,10 +7,10 @@
 typedef struct Data Data;
 
 // Constrói um elemento Data
-Data *data_construct(float x, float y, float demand);
+Data *data_construct(float x, float y, int demand);
 
 // Atribui uma demanda para um elemento Data
-void data_set_demand(Data *d, float demand);
+void data_set_demand(Data *d, int demand);
 
 // Retorna o atributo x de um elemento Data
 float data_return_x(Data *d);
@@ -19,7 +19,7 @@ float data_return_x(Data *d);
 float data_return_y(Data *d);
 
 // Retorna o atributo demand de um elemento Data
-float data_return_demand(Data *d);
+int data_return_demand(Data *d);
 
 // Imprime os atributos de um elemento Data
 void data_print(Data *d);
